+++ packages/commander-tui/src/factories/index.ts
@@ -1,1 +1,2 @@
+export * from './isomorphic-logger.factory.js'
 export * from './screen.factory.ts'